requirements.txt
rasa-core==0.10.3
rasa-nlu[spacy]==0.13.0
absl-py==0.4.1
APScheduler==3.5.3
asn1crypto==0.24.0
astor==0.7.1
attrs==18.2.0
Automat==0.7.0
autopep8==1.4
boto3==1.8.7
botocore==1.11.7
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
click==6.7
cloudpickle==0.5.5
coloredlogs==10.0
colorhash==1.0.2
ConfigArgParse==0.13.0
constantly==15.1.0
cryptography==2.3.1
cssselect==1.0.3
cycler==0.10.0
cymem==1.31.2
cytoolz==0.9.0.1
decorator==4.3.0
dill==0.2.8.2
distro==1.3.0
docopt==0.6.2
docutils==0.14
fakeredis==0.10.3
fbmessenger==5.3.0
Flask==1.0.2
Flask-Cors==3.0.6
future==0.16.0
gast==0.2.0
gevent==1.3.6
graphviz==0.8.4
greenlet==0.4.14
grpcio==1.14.2
h5py==2.8.0
humanfriendly==4.16.1
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
itsdangerous==0.24
Jinja2==2.10
jmespath==0.9.3
jsonpickle==0.9.6
jsonschema==2.6.0
Keras==2.2.2
Keras-Applications==1.0.4
Keras-Preprocessing==1.0.2
kiwisolver==1.0.1
klein==17.10.0
lxml==4.2.4
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.2.3
mattermostwrapper==2.1
msgpack==0.5.6
msgpack-numpy==0.4.3.1
murmurhash==0.28.0
networkx==2.1
numpy==1.15.1
packaging==17.1
pandas==0.23.4
parsel==1.5.0
pathlib==1.0.1
pdfminer2==20151206
pdfx==1.3.0
pika==0.11.2
plac==0.9.6
preshed==1.0.1
protobuf==3.6.1
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycodestyle==2.4.0
pycparser==2.18
PyDispatcher==2.0.5
PyHamcrest==1.9.0
PyJWT==1.6.4
pykwalify==1.6.0
pymongo==3.7.1
pyOpenSSL==18.0.0
pyparsing==2.2.0
PySocks==1.6.8
python-crfsuite==0.9.6
python-dateutil==2.7.3
python-telegram-bot==10.1.0
pytz==2018.5
PyYAML==3.13
queuelib==1.5.0
redis==2.10.6
regex==2017.4.5
requests==2.19.1
rocketchat-API==0.6.22
ruamel.yaml==0.15.64
s3transfer==0.1.13
scikit-learn==0.19.2
scipy==1.1.0
Scrapy==1.5.1
service-identity==17.0.0
simplejson==3.16.0
six==1.11.0
sklearn-crfsuite==0.3.6
slackclient==1.2.1
spacy==2.0.12
tabula-py==1.2.0
tabulate==0.8.2
tensorboard==1.10.0
tensorflow==1.10.1
termcolor==1.1.0
thinc==6.10.3
toolz==0.9.0
tqdm==4.25.0
twilio==6.16.4
Twisted==18.7.0
typing==3.6.6
tzlocal==1.5.1
ujson==1.35
urllib3==1.23
w3lib==1.19.0
websocket-client==0.52.0
Werkzeug==0.14.1
wrapt==1.10.11
zope.interface==4.5.0
python-crontab
